chore: Renaming authHeader
