+++ src/api/server/ajaxRouter.ts
@@ -4,7 +4,7 @@ import express from "express"
 import handlebars from "handlebars"
 import jwt from "jsonwebtoken"
 import { ObjectID } from "mongodb"
-import AuthHeader from "./lib/auth-header"
+import AuthHeader from "./lib/authHeader"
 import mailer from "./lib/mailer"
 import { db } from "./lib/mongo"
 import serverSettings from "./lib/settings"
