+++ src/api/server/lib/authHeader.ts
